Update final project instructions file
There's one gotcha here, where I needed to change something in
feature_format.py. If you pickle/unpickle a dictionary, Python may
iterate over it in a different order. That means that unpacking the
dataset into feature/label vectors may happen in a different order, so
different observations will end up in different folds of the cross
validation. That means that the results of running the testing script
with the live version of the dataset is different from the
pickled/unpickled version. We now sort the keys before iterating over
them to ensure a consistent ordering.