Correct feature name
In the original file, the changed lines refer to the 'MDEV' feature.

However, there is no such feature in the dataset (https://archive.ics.uci.edu/ml/datasets/Housing).  There is, however, a 'MEDV' feature, and this is likely what the changed lines are referring to.

Accordingly, change all mention of 'MDEV' to 'MEDV' for consistency with the original dataset.