Correct feature name
The original CSV's final column reads 'MDEV'.

However, in the original dataset, there is no such feature 'MDEV'.  There is, however, a feature named 'MEDV', and it is likely that this feature was intended in this CSV.

Accordingly, change the header from 'MDEV' to 'MEDV' for consistency with the original dataset.